+++ src/main/java/com/tickatch/reservationseatservice/reservationseat/domain/exception/ReservationSeatException.java
@@ -1,4 +1,4 @@
-package com.tickatch.reservationseatservice.reservationseat.domain;
+package com.tickatch.reservationseatservice.reservationseat.domain.exception;
 
 import io.github.tickatch.common.error.BusinessException;
 import io.github.tickatch.common.error.ErrorCode;
