+++ src/test/java/com/tickatch/reservationseatservice/reservationseat/domain/PriceTest.java
@@ -3,6 +3,7 @@ package com.tickatch.reservationseatservice.reservationseat.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.tickatch.reservationseatservice.reservationseat.domain.vo.Price;
 import org.junit.jupiter.api.Test;
 
 class PriceTest {
